Move fee payment computations to node-common
